requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
git-filter-repo==2.47.0
idna==3.10
imbalanced-learn==0.14.0
joblib==1.4.2
kombu==5.5.4
numpy==2.2.4
opencv-contrib-python==4.12.0.88
packaging==25.0
pandas==2.2.3
prompt_toolkit==3.0.51
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
redis>=5.0,<6.0
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
xgboost==3.0.0
dj-database-url>=2.2
whitenoise>=6.7
gunicorn==23.0.0